Move markdown translator to its own file
